--- internal/model.go
@@ -1,6 +1,0 @@
-package internal
-
-type URL struct {
-	Full  string
-	Short string
-}
