+++ pkg/shortener/repository.go
@@ -1,6 +1,13 @@
-package internal
+package shortener
 
-import "fmt"
+import (
+	"errors"
+	"fmt"
+)
+
+var (
+	ErrorURLNotFound = errors.New("url not found")
+)
 
 type Repository map[string]string
 
@@ -15,12 +22,13 @@ func (r Repository) Save(url *URL) {
 func (r Repository) FindByShort(short string) (url *URL, err error) {
 	full, exists := r[short]
 	if !exists {
-		return nil, fmt.Errorf("url not found for %q", short)
+		return nil, fmt.Errorf("%w for %q", ErrorURLNotFound, short)
 	}
 
 	url = &URL{
 		Full:  full,
 		Short: short,
 	}
+
 	return url, nil
 }
