+++ pkg/shortener/shortener_test/use_case_test.go
@@ -1,9 +1,11 @@
-package internal
+package shortener_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/cauchy2384/go-url-shortener/pkg/shortener"
 )
 
 const (
@@ -12,7 +14,7 @@ const (
 )
 
 func TestShortenURL(t *testing.T) {
-	s := NewURLShortener()
+	s := shortener.NewService()
 
 	short, err := s.ShortenURL(mediumURL)
 	assert.NoError(t, err)
@@ -20,7 +22,7 @@ func TestShortenURL(t *testing.T) {
 }
 
 func TestShortenURL_SameURLs(t *testing.T) {
-	s := NewURLShortener()
+	s := shortener.NewService()
 
 	short1, err := s.ShortenURL(mediumURL)
 	assert.NoError(t, err)
@@ -32,7 +34,7 @@ func TestShortenURL_SameURLs(t *testing.T) {
 }
 
 func TestShortenURL_DiffURLs(t *testing.T) {
-	s := NewURLShortener()
+	s := shortener.NewService()
 
 	short1, err := s.ShortenURL(mediumURL)
 	assert.NoError(t, err)
@@ -44,7 +46,7 @@ func TestShortenURL_DiffURLs(t *testing.T) {
 }
 
 func TestRetrieveURL_Exists(t *testing.T) {
-	s := NewURLShortener()
+	s := shortener.NewService()
 
 	short, err := s.ShortenURL(mediumURL)
 	assert.NoError(t, err)
@@ -56,10 +58,9 @@ func TestRetrieveURL_Exists(t *testing.T) {
 }
 
 func TestRetrieveURL_NotExists(t *testing.T) {
-	s := NewURLShortener()
+	s := shortener.NewService()
 
-	short, err := shorten(mediumURL)
-	assert.NoError(t, err)
+	short := "SMTH"
 
 	url, err := s.RetrieveURL(short)
 	assert.Error(t, err)
